feat: Per-provider `KineticLib.AuthTokenStore`
Increases the responsiveness of `AuthTokenStore`. Currently, there is
one `AuthTokenStore` for all token providers and credentials
combinations. This can lead to cascading `GenServer` call timeouts.

As an example, assume two different token providers:

- A, with an average 2 second response time
- B, with an average 7 second response time

The default request timeout is 5 seconds. Because we know that B takes
longer, its requests always have a timeout value of 8 seconds.

If both A and B need refreshing at the same time, then we have the
following possible negative cases:

- (A then B): A succeeds in 2 seconds, but B times out because the time
  it spent _waiting_ for A to complete is counted in its time elapsed.
- (B then A): B succeeds in 6 seconds, but A times out because 6 seconds
  is longer than the default 5 second timeout.

This change moves the provider retrieval logic from `AuthTokenStore` to
`AuthTokenStore.ProviderStore`. The delegating functions in
`AuthTokenStore` are used to start a per-provider named
`AuthTokenSTore.ProviderStore`, which means that only requests for the
_same_ provider can result in a possible timeout chain.

That is:

- If a request is made for A, the request is made to the ProviderStore
  instance _for_ A.
- If a request is made for B, the request is made to the ProviderStore
  instance _for_ B.

It is still possible for two requests to B to time out in this scenario,
   especially if they are for different credentials, but this can be
   mitigated by implementing an extended timeout in the `TokenProvider`.

In cases where this happens frequently, the recommendation is to create
a secondary provider that delegates token retrieval calls to the
canonical implementation.